Move the changes api to its own controller
